+++ views/settings/form.php
@@ -7,21 +7,24 @@
 
 use humhub\modules\questions\models\forms\ContainerSettings;
 use humhub\modules\space\models\Space;
-use humhub\widgets\Button;
-use yii\bootstrap\ActiveForm;
+use humhub\widgets\bootstrap\Button;
+use humhub\widgets\form\ActiveForm;
 
 /* @var ContainerSettings $settings */
 ?>
 <div class="panel panel-default">
+
     <div class="panel-heading"><?= Yii::t('QuestionsModule.base', '<strong>Questions</strong> settings') ?></div>
+
     <div class="panel-body">
-        <div class="help-block">
+        <?php $form = ActiveForm::begin() ?>
+        <div class="form-text">
             <?= $settings->contentContainer instanceof Space
                 ? Yii::t('QuestionsModule.base', 'Settings of the "Questions" module for this single Space.')
                 : Yii::t('QuestionsModule.base', 'Settings of the "Questions" module for your Profile.')?>
         </div>
-        <br>
-        <?php $form = ActiveForm::begin() ?>
+
+        <hr>
 
         <?= $form->field($settings, 'showAnswersInStream')->checkbox() ?>