+++ widgets/AnswersToggleButton.php
@@ -9,7 +9,7 @@
 namespace humhub\modules\questions\widgets;
 
 use humhub\components\Widget;
-use humhub\widgets\Button;
+use humhub\widgets\bootstrap\Button;
 use Yii;
 
 class AnswersToggleButton extends Widget
@@ -34,7 +34,7 @@ class AnswersToggleButton extends Widget
 
     private function renderCollapseButton(): string
     {
-        $button = Button::info(Yii::t('QuestionsModule.base', 'Collapse all answers ({count})', [
+        $button = Button::accent(Yii::t('QuestionsModule.base', 'Collapse all answers ({count})', [
             'count' => '<span class="questions-answers-count">' . $this->count . '</span>',
         ]))
             ->icon('arrow-up')
@@ -43,7 +43,7 @@ class AnswersToggleButton extends Widget
             ->loader(false);
 
         if (!$this->isVisibleButton('collapse')) {
-            $button->style('display:none');
+            $button->cssClass('d-none');
         }
 
         return $button;
@@ -51,7 +51,7 @@ class AnswersToggleButton extends Widget
 
     private function renderExpandButton(): string
     {
-        $button = Button::info(Yii::t('QuestionsModule.base', 'Expand all answers ({count})', [
+        $button = Button::accent(Yii::t('QuestionsModule.base', 'Expand all answers ({count})', [
             'count' => '<span class="questions-answers-count">' . $this->count . '</span>',
         ]))
             ->icon('arrow-down')
@@ -60,7 +60,7 @@ class AnswersToggleButton extends Widget
             ->loader(false);
 
         if (!$this->isVisibleButton('expand')) {
-            $button->style('display:none');
+            $button->cssClass('d-none');
         }
 
         return $button;